extracted shared controls and markdown rendering
