+++ frontend/src/components/IconButton.jsx
@@ -1,0 +1,43 @@
+import { cx, CONTROL_MOTION, SOFT_SURFACE } from "../uiShared.js";
+
+export function IconButton({ label, children, className, ...props }) {
+  return (
+    <button
+      aria-label={label}
+      title={label}
+      className={cx(
+        "inline-flex h-10 w-10 items-center justify-center rounded-full border border-line bg-white/[0.04] text-neutral-300 focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/50",
+        CONTROL_MOTION,
+        SOFT_SURFACE,
+        "hover:border-white/15 hover:bg-white/[0.075] hover:text-white",
+        className,
+      )}
+      {...props}
+    >
+      {children}
+    </button>
+  );
+}
+
+export function MaskIcon({ src, size = 19, className }) {
+  const maskStyle = {
+    width: size,
+    height: size,
+    maskImage: `url("${src}")`,
+    WebkitMaskImage: `url("${src}")`,
+    maskRepeat: "no-repeat",
+    WebkitMaskRepeat: "no-repeat",
+    maskPosition: "center",
+    WebkitMaskPosition: "center",
+    maskSize: "contain",
+    WebkitMaskSize: "contain",
+  };
+
+  return (
+    <span
+      aria-hidden="true"
+      className={cx("inline-block shrink-0 bg-current", className)}
+      style={maskStyle}
+    />
+  );
+}
